refactor(cli): extract clap Cli/Commands + run_cli dispatch into cli.rs (#233)
Move the clap `Cli` / `Commands` definitions, the `CLI_STYLES` const, and the
`run()` / `run_cli()` dispatch out of the ~4.4k-line lib.rs into a new
src/cli.rs. lib.rs drops to ~3.8k lines (leftover helpers + tests).

- `pub use crate::cli::run;` keeps `rvpm::run()` and `src/main.rs`
  (`fn main() { rvpm::run() }`) unchanged.
- `Cli` is `pub(crate)` and re-exported at the crate root so
  `commands::run_completion`'s `<Cli as CommandFactory>` and the
  completion-generation unit test keep resolving it via `use crate::*`.
- Dropped now-unused lib.rs imports (`clap::{Parser, Subcommand}`,
  `commands::*`, `self_update::*`); cli.rs carries them instead.
- Sync the `commands/mod.rs` module doc to point at `cli.rs::run_cli` and the
  new `Cli` / `Commands` location (was stale after the move).

Behaviour-preserving: code moved verbatim. cargo make check green: 822 tests,
plural doctest preserved, clippy -D warnings --all-targets, fmt, lock-check.

The leftover-helper sweep (moving sole-caller helpers from lib.rs into their
command modules) is left as a follow-up; genuinely shared helpers stay in
lib.rs per the issue.

Refs #233, #217.
